requirements.txt
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==2.0.5
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyftdi==0.55.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
pyusb==1.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
six==1.17.0
sysv-ipc==1.1.0
typing-extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.21.0
